+++ src/backend/marsha/markdown/api.py
@@ -1,9 +1,10 @@
 """Declare API endpoints with Django RestFramework viewsets."""
 from django.conf import settings
+from django.db.models import Q
 from django.utils import timezone
 
 import django_filters
-from rest_framework import mixins, viewsets
+from rest_framework import filters, mixins, viewsets
 from rest_framework.decorators import action
 from rest_framework.response import Response
 from rest_framework.status import HTTP_200_OK, HTTP_400_BAD_REQUEST
@@ -37,14 +38,26 @@ class MarkdownDocumentViewSet(
     APIViewMixin,
     ObjectPkMixin,
     mixins.CreateModelMixin,
+    mixins.ListModelMixin,
     mixins.RetrieveModelMixin,
     mixins.UpdateModelMixin,
     viewsets.GenericViewSet,
 ):
     """Viewset for the API of the Markdown document object."""
 
-    queryset = MarkdownDocument.objects.prefetch_related("translations").all()
+    queryset = (
+        MarkdownDocument.objects.prefetch_related("translations")
+        .select_related("playlist")
+        .all()
+    )
     serializer_class = serializers.MarkdownDocumentSerializer
+    filter_backends = [
+        filters.OrderingFilter,
+        django_filters.rest_framework.DjangoFilterBackend,
+    ]
+    ordering_fields = ["created_on"]
+    ordering = ["-created_on"]
+    filterset_class = MarkdownDocumentFilter
 
     permission_classes = [
         (
@@ -83,6 +96,32 @@ class MarkdownDocumentViewSet(
             permission_classes = self.permission_classes
         return [permission() for permission in permission_classes]
 
+    def _get_list_queryset(self):
+        """Build the queryset used on the list action."""
+        queryset = (
+            super()
+            .get_queryset()
+            .filter(
+                Q(
+                    playlist__organization__user_accesses__user_id=self.request.user.id,
+                    playlist__organization__user_accesses__role=ADMINISTRATOR,
+                )
+                | Q(
+                    playlist__user_accesses__user_id=self.request.user.id,
+                    playlist__user_accesses__role=ADMINISTRATOR,
+                )
+            )
+        )
+
+        return queryset
+
+    def get_queryset(self):
+        """Redefine the queryset to use based on the current action."""
+        if self.action in ["list"]:
+            return self._get_list_queryset()
+
+        return super().get_queryset()
+
     def create(self, request, *args, **kwargs):
         """Create one document based on the request payload."""
         try:
@@ -94,25 +133,6 @@ class MarkdownDocumentViewSet(
         serializer = self.get_serializer(document)
 
         return Response(serializer.data, status=201)
-
-    def list(self, request, *args, **kwargs):
-        """List markdown documents belonging to user organizations."""
-        queryset = self.get_queryset().filter(
-            playlist__organization__users__id=request.user.id,
-            playlist__organization__user_accesses__role=ADMINISTRATOR,
-        )
-        filter_set = MarkdownDocumentFilter(request.query_params, queryset=queryset)
-        page = self.paginate_queryset(
-            self.filter_queryset(
-                filter_set.qs.order_by(
-                    "-created_on",
-                    "-playlist__created_on",
-                    "-playlist__organization__created_on",
-                )
-            )
-        )
-        serializer = self.get_serializer(page, many=True)
-        return self.get_paginated_response(serializer.data)
 
     @action(
         methods=["get"],